moved configuration scripts
