+++ Assets/Scripts/Tools/ToolSpawner.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Netherlands3D.Twin
