+++ Assets/_Application/Configuration/FunctionalitiesPane.cs
