+++ Assets/_Application/Configuration/SceneSwitcher.cs
@@ -1,9 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Configuration
 {
     public class SceneSwitcher : MonoBehaviour
     {
